# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - AdminRole
    - UserRole
  entity.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
  entity.TranslationHistory:
    properties:
      history:
        items:
          $ref: '#/definitions/entity.Translation'
        type: array
    type: object
...
  request.Translate:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
    required:
    - destination
    - original
    - source
    type: object
  request.UserLogin:
    properties:
      email:
        example: minhhoccode111@gmail.com
        type: string
      password:
        example: P@ssw0rd
        type: string
    required:
    - email
    - password
    type: object
  request.UserLoginRequest:
    properties:
      user:
        $ref: '#/definitions/request.UserLogin'
    type: object
  request.UserRegister:
    properties:
      email:
        example: minhhoccode111@gmail.com
        type: string
      password:
        example: P@ssw0rd
        maxLength: 50
        minLength: 8
        type: string
      username:
        example: minhhoccode111
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - password
    - username
    type: object
  request.UserRegisterRequest:
    properties:
      user:
        $ref: '#/definitions/request.UserRegister'
    type: object
  request.UserUpdate:
    properties:
      bio:
        example: Trust the process
        maxLength: 255
        type: string
      email:
        example: minhhoccode111@gmail.com
        maxLength: 320
        minLength: 5
        type: string
      image:
        example: https://www.w3schools.com/howto/img_avatar.png
        maxLength: 2048
        type: string
      password:
        example: P@ssw0rd
        maxLength: 50
        minLength: 8
        type: string
      username:
        example: minhhoccode111
        maxLength: 50
        minLength: 2
        type: string
    type: object
  request.UserUpdateRequest:
    properties:
      user:
        $ref: '#/definitions/request.UserUpdate'
    type: object
...
  response.Error:
    properties:
      error:
        example: message
        type: string
    type: object
...
  response.UserAuth:
    properties:
      bio:
        type: string
      email:
        type: string
      image:
        type: string
      role:
        $ref: '#/definitions/entity.Role'
      token:
        type: string
      username:
        type: string
    type: object
  response.UserAuthResponse:
    properties:
      user:
        $ref: '#/definitions/response.UserAuth'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Realworld API using Golang + Gin + Clean Architecture
  title: Realworld Fiber Clean API
  version: "1.0"
paths:
...
  /translation/do-translate:
    post:
      consumes:
      - application/json
      description: Translate a text
      operationId: do-translate
      parameters:
      - description: Set up translation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.Translate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Translation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Translate
      tags:
      - translation
  /translation/history:
    get:
      consumes:
      - application/json
      description: Show all translation history
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.TranslationHistory'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Show history
      tags:
      - translation
  /user:
    get:
      description: Get current User
      operationId: users-current
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Get current User
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update User
      operationId: users-update
      parameters:
      - description: Update User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - users
  /users:
    post:
      consumes:
      - application/json
      description: Register User
      operationId: users-register
      parameters:
      - description: Register User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.UserAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Register User
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login User
      operationId: users-login
      parameters:
      - description: Login User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      summary: Login User
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Token" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"